gelo_ai/requirements.txt
fastapi
uvicorn
pydantic
httpx
pillow
numpy
torch
torchvision
